Only register toolchains from top module
Because toolchain registration is global, the top level module should control this. Otherwise if one module wants to use a rust toolchain for development and testing of that module, it will also get registered when used as a dependency. This will result in a lot of rust versions being registered, but the usage of that rust version would not even be encapsulated within that module.